refactor: refactor codebase
